Added suggestion under Legal for Web design contracts
As a web design freelancer or agency, contracts are essential to avoid too many revisions and to have commitment from both the client and the designer. The web design contracts bundle is a download of 17 web design contracts, including the Andy Clarke killer contract to help web designers get started with their contracts.